chore: clean the code
